+++ semcore/tag/__tests__/index.test.tsx
@@ -163,7 +163,7 @@ describe('Tag', () => {
       </Tag>,
     );
 
-    fireEvent.keyDown(getByTestId('close'), { key: 'Enter' });
+    fireEvent.keyDown(getByTestId('close'), { code: 'Enter' });
     expect(onClick).toHaveBeenCalledTimes(1);
   });
 
@@ -177,7 +177,7 @@ describe('Tag', () => {
       </Tag>,
     );
 
-    fireEvent.keyDown(getByTestId('close'), { key: 'Enter' });
+    fireEvent.keyDown(getByTestId('close'), { code: 'Enter' });
     expect(onClick).toHaveBeenCalledTimes(0);
   });