+++ semcore/tag/src/Tag.jsx
@@ -64,6 +64,7 @@ class RootTag extends Component {
     switch (e.key) {
       case ' ':
       case 'Enter':
+        e.preventDefault();
         this.asProps.onClick?.(e);
         break;
     }
@@ -119,8 +120,9 @@ function Close(props) {
       return props.onKeyDown(event);
     }
 
-    if (event.key === 'Enter') {
-      props.onClick?.(event);
+    if (props.onClick && (event.code === 'Enter' || event.code === 'Space')) {
+      event.preventDefault();
+      props.onClick(event);
     }
   }