+++ semcore/inline-input/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [4.25.0] - 2024-02-13
+
+### Fixed
+
+- Confirm and cancel controls were ignoring `Space` click.
+
 ## [4.24.2] - 2024-02-09
 
 ### Changed