+++ semcore/switch/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [5.19.0] - 2024-02-13
+
+### Changed
+
+- Switches are clickable by `Space` now (along with `Enter` as before).
+
 ## [5.18.2] - 2024-02-09
 
 ### Changed